Refactor imagewidget plugin
